Issue #1360. Move get_form into form module to prevent circular import.
